DTO refactor, file storage

+++ be/src/main/java/com/lms/utils/Constants.java
@@ -1,5 +1,5 @@
 package com.lms.utils;
 
 public class Constants {
-    public static final int DEFAULT_DAYS_OFF = 12;
+    public static final float DEFAULT_DAYS_OFF = 12;
 }
